move spritelab unit tests
