+++ apps/test/unit/p5lab/spritelab/actionCommandsTest.js
@@ -1,8 +1,8 @@
 /* global p5 */
-import {expect} from '../../util/reconfiguredChai';
+import {expect} from '../../../util/reconfiguredChai';
 import {commands} from '@cdo/apps/p5lab/spritelab/commands/actionCommands';
 import {commands as spriteCommands} from '@cdo/apps/p5lab/spritelab/commands/spriteCommands';
-import createP5Wrapper from '../../util/gamelab/TestableP5Wrapper';
+import createP5Wrapper from '../../../util/gamelab/TestableP5Wrapper';
 
 describe('Action Commands', () => {
   let p5Wrapper, makeSprite;
