+++ apps/test/unit/p5lab/spritelab/blocksTest.js
@@ -1,4 +1,4 @@
-import {expect} from '../../util/deprecatedChai';
+import {expect} from '../../../util/deprecatedChai';
 import {sprites} from '@cdo/apps/p5lab/spritelab/blocks';
 import {
   registerReducers,
