+++ apps/test/unit/p5lab/spritelab/locationCommandsTest.js
@@ -1,7 +1,7 @@
-import {expect} from '../../util/reconfiguredChai';
+import {expect} from '../../../util/reconfiguredChai';
 import {commands} from '@cdo/apps/p5lab/spritelab/commands/locationCommands';
 import {commands as spriteCommands} from '@cdo/apps/p5lab/spritelab/commands/spriteCommands';
-import createP5Wrapper from '../../util/gamelab/TestableP5Wrapper';
+import createP5Wrapper from '../../../util/gamelab/TestableP5Wrapper';
 
 describe('Location Commands', () => {
   let p5Wrapper, makeSprite;
